Refactor DAO into separate managers. Add initial authorization manager, with dummy DB implementation.
